Add response_type to SlackApiPostWebhookMessageRequest.
As far as I can tell, actual webhooks don't have a response_type,
or an original message to delete or replace, so this type should
probably eventually be split into "PostWebhookMessage" and
"PostResponseMessage" or similar.

In the meantime, however, this is the type you use when responding
to a slash command, so pending a major semver bump, this appears
to be where this field belongs.